requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
azure-core==1.29.7
azure-iot-hub==2.6.1
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
gunicorn==21.2.0
h11==0.14.0
idna==3.6
isodate==0.6.1
Mako==1.3.0
MarkupSafe==2.1.3
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
paho-mqtt==1.6.1
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.32.0.post1
typing_extensions==4.9.0
uamqp==1.6.7
urllib3==2.1.0
uvicorn==0.25.0